# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AffordEnum": {
        "description": "PLAN \u00a75.8's documented outcome classes (\"SURVIVABLE ... STRAINED ...\nDEFICIT\"), used by both `shock_scenarios.affordability_status` and\n`offer_assessments.affordability_status` (both named `afford_enum` in\n\u00a711.3 \u2014 one shared type, not two). Added Sprint 5/T5.1/T5.3.",
        "enum": [
          "SURVIVABLE",
          "STRAINED",
          "DEFICIT"
        ],
        "title": "AffordEnum",
        "type": "string"
      },
      "AmortizationEnum": {
        "description": "PLAN \u00a75.7 \"amortisation method (`FLAT | REDUCING_BALANCE |\nFIXED_SCHEDULE`)\" \u2014 not enumerated in Appendix A. `FIXED_SCHEDULE` is\nreserved for a future lender-supplied schedule that doesn't follow either\nformula (not produced by `app/engines/loan_math.py` this sprint). Added\nSprint 5/T5.3.",
        "enum": [
          "FLAT",
          "REDUCING_BALANCE",
          "FIXED_SCHEDULE"
        ],
        "title": "AmortizationEnum",
        "type": "string"
      },
...
      "BandEnum": {
        "description": "PLAN Appendix A `band_enum` \u2014 used by `document_verification_results\n.confidence_band` (\u00a75.2: `HIGH >= 80`, `MEDIUM 50-79`, `LOW < 50`).",
        "enum": [
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "BandEnum",
        "type": "string"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "correlation_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "EssentialExpenseCoverageResponse": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "ratio": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Ratio",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "ratio"
        ],
        "title": "EssentialExpenseCoverageResponse",
        "type": "object"
      },
...
      "FreqEnum": {
        "description": "PLAN Appendix A `freq_enum`.",
        "enum": [
          "MONTHLY",
          "BIWEEKLY",
          "WEEKLY"
        ],
        "title": "FreqEnum",
        "type": "string"
      },
...
      "LatePenaltyTermsResponse": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "basis": {
            "enum": [
              "OVERDUE_INSTALMENT_PER_DAY",
              "OVERDUE_INSTALMENT_PER_MONTH",
              "FIXED"
            ],
            "title": "Basis",
            "type": "string"
          },
          "rate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          },
          "trigger_days": {
            "minimum": 0.0,
            "title": "Trigger Days",
            "type": "integer"
          }
        },
        "required": [
          "trigger_days",
          "rate",
          "amount",
          "basis"
        ],
        "title": "LatePenaltyTermsResponse",
        "type": "object"
      },
      "LenderResponse": {
        "properties": {
          "lender_id": {
            "format": "uuid",
            "title": "Lender Id",
            "type": "string"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "regulatory_status": {
            "$ref": "#/components/schemas/RegStatusEnum"
          }
        },
        "required": [
          "lender_id",
          "name",
          "regulatory_status",
          "logo_url"
        ],
        "title": "LenderResponse",
        "type": "object"
      },
...
      "OfferRatingEnum": {
        "description": "Gap-fill (PLAN \u00a724.11): PLAN \u00a711.3's `offer_assessments` row types\n`total_cost_status`/`timing_status` as a generic `status_enum` without\nnaming its members. A three-tier rating (rather than reusing `band_enum`,\nwhich is the Trust-Layer-specific HIGH/MEDIUM/LOW vocabulary) keeps offer\ncopy in its own domain language. Added Sprint 5/T5.3.",
        "enum": [
          "GOOD",
          "FAIR",
          "POOR"
        ],
        "title": "OfferRatingEnum",
        "type": "string"
      },
      "OfferResponse": {
        "properties": {
          "admin_fee": {
            "title": "Admin Fee",
            "type": "integer"
          },
          "affordability_status": {
            "$ref": "#/components/schemas/AffordEnum"
          },
          "amortization_method": {
            "$ref": "#/components/schemas/AmortizationEnum"
          },
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "due_date": {
            "title": "Due Date",
            "type": "integer"
          },
          "effective_annual_rate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Annual Rate"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "financed_fee": {
            "title": "Financed Fee",
            "type": "integer"
          },
          "frequency": {
            "$ref": "#/components/schemas/FreqEnum"
          },
          "instalment_amount": {
            "title": "Instalment Amount",
            "type": "integer"
          },
          "interest_amount": {
            "title": "Interest Amount",
            "type": "integer"
          },
          "late_penalty_terms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatePenaltyTermsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "lender": {
            "$ref": "#/components/schemas/LenderResponse"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "net_disbursed_amount": {
            "title": "Net Disbursed Amount",
            "type": "integer"
          },
          "nominal_rate": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nominal Rate"
          },
          "nominal_rate_basis": {
            "const": "ANNUAL_NOMINAL",
            "title": "Nominal Rate Basis",
            "type": "string"
          },
          "offer_id": {
            "format": "uuid",
            "title": "Offer Id",
            "type": "string"
          },
          "offer_source": {
            "$ref": "#/components/schemas/OfferSourceEnum"
          },
          "payment_schedule": {
            "items": {
              "$ref": "#/components/schemas/PaymentScheduleEntryResponse"
            },
            "title": "Payment Schedule",
            "type": "array"
          },
          "principal_amount": {
            "title": "Principal Amount",
            "type": "integer"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/ReasonCodeResponse"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "refinancing_dependency": {
            "title": "Refinancing Dependency",
            "type": "boolean"
          },
          "remaining_essential_expense_coverage": {
            "$ref": "#/components/schemas/EssentialExpenseCoverageResponse"
          },
          "safe_offer_score": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Safe Offer Score",
            "type": "string"
          },
          "safety_band": {
            "$ref": "#/components/schemas/OfferSafetyBandEnum"
          },
          "service_fee": {
            "title": "Service Fee",
            "type": "integer"
          },
          "shock_resilience_status": {
            "$ref": "#/components/schemas/BandEnum"
          },
          "simulation_notice": {
            "anyOf": [
              {
                "const": "SIMULATED offer for comparison only; not a real provider endorsement.",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Notice"
          },
          "tenor_months": {
            "title": "Tenor Months",
            "type": "integer"
          },
          "timing_status": {
            "$ref": "#/components/schemas/OfferRatingEnum"
          },
          "total_cost_status": {
            "$ref": "#/components/schemas/OfferRatingEnum"
          },
          "total_repayment": {
            "title": "Total Repayment",
            "type": "integer"
          },
          "upfront_fee": {
            "title": "Upfront Fee",
            "type": "integer"
          },
          "warning_flags": {
            "items": {
              "type": "string"
            },
            "title": "Warning Flags",
            "type": "array"
          }
        },
        "required": [
          "offer_id",
          "lender",
          "offer_source",
          "principal_amount",
          "net_disbursed_amount",
          "instalment_amount",
          "tenor_months",
          "amortization_method",
          "nominal_rate",
          "nominal_rate_basis",
          "effective_annual_rate",
          "interest_amount",
          "upfront_fee",
          "financed_fee",
          "service_fee",
          "admin_fee",
          "total_repayment",
          "late_penalty_terms",
          "payment_schedule",
          "due_date",
          "frequency",
          "safe_offer_score",
          "safety_band",
          "rank",
          "affordability_status",
          "shock_resilience_status",
          "total_cost_status",
          "timing_status",
          "warning_flags",
          "refinancing_dependency",
          "remaining_essential_expense_coverage",
          "reason_codes",
          "explanation",
          "model_version",
          "config_hash",
          "simulation_notice"
        ],
        "title": "OfferResponse",
        "type": "object"
      },
      "OfferSafetyBandEnum": {
        "description": "PLAN \u00a75.9: \"Bands: SAFE >= 75, CAUTION 50-74, UNSAFE < 50\" \u2014 named\nonly in prose; `offer_assessments` stores `safe_offer_score` (`NUMERIC`)\nand `rank` (`INT`) but no band column, so this is computed on read (same\npattern as `ShockResilienceBandEnum` above). Added Sprint 5/T5.3.",
        "enum": [
          "SAFE",
          "CAUTION",
          "UNSAFE"
        ],
        "title": "OfferSafetyBandEnum",
        "type": "string"
      },
      "OfferSourceEnum": {
        "description": "PLAN \u00a711.3 `lender_offers.offer_source` \u2014 member set given inline\n(`SIMULATED, LENDER_API, MANUAL_LENDER_ENTRY`), not in Appendix A. Added\nSprint 5/T5.3.",
        "enum": [
          "SIMULATED",
          "LENDER_API",
          "MANUAL_LENDER_ENTRY"
        ],
        "title": "OfferSourceEnum",
        "type": "string"
      },
      "OffersListResponse": {
        "properties": {
          "assessment_id": {
            "format": "uuid",
            "title": "Assessment Id",
            "type": "string"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/OfferResponse"
            },
            "title": "Offers",
            "type": "array"
          }
        },
        "required": [
          "assessment_id",
          "offers"
        ],
        "title": "OffersListResponse",
        "type": "object"
      },
      "PaymentScheduleEntryResponse": {
        "properties": {
          "interest_component": {
            "title": "Interest Component",
            "type": "integer"
          },
          "payment_amount": {
            "title": "Payment Amount",
            "type": "integer"
          },
          "period": {
            "title": "Period",
            "type": "integer"
          },
          "principal_component": {
            "title": "Principal Component",
            "type": "integer"
          },
          "remaining_balance": {
            "title": "Remaining Balance",
            "type": "integer"
          }
        },
        "required": [
          "period",
          "payment_amount",
          "principal_component",
          "interest_component",
          "remaining_balance"
        ],
        "title": "PaymentScheduleEntryResponse",
        "type": "object"
      },
...
      "ReasonCodeResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "required": [
          "code",
          "description"
        ],
        "title": "ReasonCodeResponse",
        "type": "object"
      },
...
      "RegStatusEnum": {
        "description": "PLAN \u00a711.3 `lenders.regulatory_status` (`reg_status_enum`) documents\ntwo members (`REGULATED, UNLISTED`); a third, `SIMULATED_REGULATED_PROVIDER`,\nis required by FR-11 AC5 for MVP seeded providers so offer copy never\nimplies a real regulated endorsement (PLAN \u00a76.4/\u00a716 non-goals: no real\nlender integration in MVP). Added Sprint 5/T5.3.",
        "enum": [
          "REGULATED",
          "UNLISTED",
          "SIMULATED_REGULATED_PROVIDER"
        ],
        "title": "RegStatusEnum",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "crediwise-backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assessments/{assessment_id}/offers": {
      "get": {
        "description": "PLAN \u00a712.2 `GET /assessments/{id}/offers` -- ranked offers (FR-11 AC1:\nranked by safety, never commission).",
        "operationId": "get_assessment_offers_api_v1_assessments__assessment_id__offers_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffersListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict or reassessment required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Stable validation error envelope"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Get Assessment Offers",
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "description": "PLAN \u00a712.2 `POST /assessments/{id}/offers` -- seed/simulate offers\n(FR-11). Deterministic offer templates against the seeded `lenders`\ncatalog; see `app/services/offer_service.py` module docstring for the\nre-seeding behaviour.",
        "operationId": "create_assessment_offers_api_v1_assessments__assessment_id__offers_post",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffersListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict or reassessment required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Stable validation error envelope"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Create Assessment Offers",
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/api/v1/offers/{offer_id}/safety": {
      "get": {
        "operationId": "get_offer_safety_api_v1_offers__offer_id__safety_get",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict or reassessment required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Stable validation error envelope"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Get Offer Safety",
        "tags": [
          "offers"
        ]
      }
    },
...
  }
}